refacto function
