+++ efrei-internship-front-ta/helper/HelpStudent.js
@@ -6,3 +6,12 @@ export const getStudentName = (idStudent, listStudent) => {
     }
     return "Unknown";
 }
+
+export const formatDate = (date) => {
+    const formattedDate = new Date(date);
+    const year = formattedDate.toLocaleString("default", {year: "numeric"});
+    const month = formattedDate.toLocaleString("default", {month: "2-digit"});
+    const day = formattedDate.toLocaleString("default", {day: "2-digit"});
+
+    return `${year}-${month}-${day}`;
+}
